chore:#39 constant와 type 파일을 컨벤션에 맞게 분리
